==> notebooks/0.2.Creaccion_columnas_metereologia copy.ipynb <==
## Importar librerías


import pandas as pd
import numpy as np
import os
from data_util import calc_missing

## Cargar datos


file_names = os.listdir('../data/metereologia/')
file_names

## Crear un único dataset con los datos del año 2022

df_meteo = pd.DataFrame()
for file_name in file_names:
    df_current = pd.read_csv('../data/metereologia/' + file_name, delimiter=';')
    print(file_name, df_current.shape)
    df_meteo = pd.concat([df_meteo, df_current], ignore_index=True)
df_meteo.head(3)

## Nos quedamos con los parámetros de las estaciones que nos interesan:

8, 18, 24, 35, 36

# df_meteo  = df_meteo[(df_meteo['ESTACION'] == 8) | (df_meteo['ESTACION'] == 18) | (df_meteo['ESTACION'] == 24 ) |
#                     (df_meteo['ESTACION'] == 35) | (df_meteo['ESTACION'] == 36)]
df_meteo  = df_meteo[(df_meteo['ESTACION'] == 8)]
df_meteo.head()

# sorted(df_meteo['ESTACION'].unique())

# Nos quedamos con los parámetros que nos interesan: 
 velocidad del viento, la temperatura, la humedad relativa y las precipitaciones.


df_meteo = df_meteo[(df_meteo['MAGNITUD'] == 81) | (df_meteo['MAGNITUD'] == 82) | (df_meteo['MAGNITUD'] == 83) | (df_meteo['MAGNITUD'] == 86) | (df_meteo['MAGNITUD'] == 89)]
df_meteo.head()

df_meteo.info()

## Preparación de los datos de salida

data_columns = [ 'H01', 'H02', 'H03', 'H04', 'H05', 'H06', 'H07', 'H08', 'H09', 'H10', 'H11', 'H12', 'H13', 'H14', 'H15', 'H16', 'H17', 'H18', 'H19', 'H20', 'H21', 'H22', 'H23', 'H24'] # feature
veri_columns = [ 'V01', 'V02', 'V03', 'V04', 'V05', 'V06', 'V07', 'V08', 'V09', 'V10', 'V11', 'V12', 'V13', 'V14', 'V15', 'V16', 'V17', 'V18', 'V19', 'V20', 'V21', 'V22', 'V23', 'V24']

hora_column = np.arange(1,25) # feature label

data = pd.DataFrame(columns=['HORA', 'DATO', 'VERIFICADO'])
aux = pd.DataFrame(columns=['ANO', 'MES', 'DIA', 'HORA', 'DATO', 'VERIFICADO', 'ESTACION']) # aux de próxima fila a ser añadida
for  (h,d,v) in zip(hora_column, data_columns, veri_columns): 
    aux
    aux['DATO'] = df_meteo[ data_columns[h-1] ] 
    aux[['CODIGO', 'MAGNITUD', 'TECNICA']] = df_meteo['PUNTO_MUESTREO'].str.split('_', n=2, expand=True)
    aux['VERIFICADO'] = df_meteo[ veri_columns[h-1] ] 
    aux['HORA'] = h-1
    aux[['ANO', 'MES', 'DIA']] = df_meteo[['ANO','MES','DIA']]
    #aux['MAGNITUD'] = df_aire ['MAGNITUD']
    aux['ESTACION'] = df_meteo ['ESTACION']

    data = pd.concat([data, aux], ignore_index=True)
data = data.astype({'HORA': 'int32'}, copy = False)

data.head()

# Añadimos columna fecha de formato datetime

data['FECHA'] = pd.to_datetime(data[['ANO','MES','DIA','HORA']].rename(columns = {'ANO': 'year', 'MES': 'month','DIA': 'day', 'HORA': 'hour'}))
data.head()

# Ordenamos datos por orden cronológico

data = data.sort_values(['FECHA','HORA'])
data.head()

## Marcar datos no válidos en el DataFrame

Datos no verificados

data.loc[data['VERIFICADO'] != 'V', ['DATO']] = np.nan
data.loc[data['VERIFICADO'] != 'V']

# Borramos columna verificado

data = data.drop('VERIFICADO', axis=1)
data.info()

## Añadir columnas que faltan y formatearlas

#data = data.set_index(['CODIGO','FECHA', 'ANO', 'MES', 'DIA', 'HORA', 'ESTACION','MAGNITUD'])['DATO'].unstack(level=-1).reset_index()
data = data.set_index(['FECHA','MAGNITUD'])['DATO'].unstack(level=-1).reset_index()

data.head()


## Renombrar columnas

#data = data.rename(columns={'81': 'VEL_VIENTO', '82': 'DIR_VIENTO', '83': 'TEMPERATURA', '86': 'HUMEDAD','89': 'PRECIPITACION'}, copy = False, errors = 'raise')
data = data.rename(columns={'83': 'TEMPERATURA', '86': 'HUMEDAD'}, copy = False, errors = 'raise')

data.head()

# Añadimos columna para el día de la semana

# dates = data['FECHA'].dt.to_pydatetime()
# days_of_week = ['Lunes', 'Martes', 'Miercoles', 'Jueves', 'Viernes', 'Sabado', 'Domingo']
# data['DIA_SEMANA'] = np.array([days_of_week[d.weekday()] for d in dates])
# data.head()

## Datos no válidos

calc_missing(data)

#variables = ['VEL_VIENTO', 'DIR_VIENTO', 'TEMPERATURA','HUMEDAD','PRECIPITACION']
variables = [ 'TEMPERATURA','HUMEDAD']
datos = data [variables]
datos

# buscamos si hay algún caso con algún campo con valor 0 o negativo donde no debería
(datos <= 0).any()

# dir_viento_neg_idx = data['DIR_VIENTO'] < 0
# data[dir_viento_neg_idx]

# precipitacion_neg_idx = data['PRECIPITACION'] < 0
# data[precipitacion_neg_idx]

## Tratamiento de valores perdidos o no válidos

# # Número de datos no válidos
# null_idx = data.isnull().any(axis=1)
# len(data[data.isnull().any(axis=1)])

# null_idx = data.isnull().any(axis=1)
# data[null_idx]

# Otra alternativa mejor es llenarlos con último valor válido anterior:
# data.fillna(method='ffill', inplace = True )

# Número de datos no válidos
# len(data[data.isnull().any(axis=1)])

# data.fillna(0, inplace=True)

# Número de datos no válidos
# len(data[data.isnull().any(axis=1)])

# data[null_idx]

data.shape

# data['DIR_VIENTO'].values

# # Discretizamos variable dirección viento en las siguientes 8 categorías separadas cada una por 45º: [338, 23, 68, 113, 158, 203, 248, 293]
# data['DIR_VIENTO'] = pd.to_numeric(data['DIR_VIENTO'], errors='coerce')

# data['DIR_VIENTO_CARDINAL'] = ((data['DIR_VIENTO'] + 22) / 45) % 8
# #data['dir_viento_cardinal'] = ((data['dir_viento'] + 22) / 45) % 8
# direcciones = ['N', 'NE', 'E', 'SE', 'S', 'SW', 'W', 'NW']

# data['DIR_VIENTO_CARDINAL'] = data['DIR_VIENTO_CARDINAL'].apply(lambda x: direcciones[int (x)])
# data['DIR_VIENTO_CARDINAL']
# data.head()

## Salvar datos preprocesados

data.info()

data.head(2)

data.isnull().sum()


data.to_csv('../data/auxiliar/meteo_2022.csv',index=False) # Salvado con todos los datos
