==> notebooks/0.2.Creaccion_columnas_metereologia.ipynb <==
## Importar librerías


import pandas as pd
import numpy as np
import os
from data_util import calc_missing

## Cargar datos


ruta = '../data/metereologia/'
ruta_nuevoCSV = '../data/auxiliar/meteo.csv'
file_names = os.listdir(ruta)
file_names

## Crear un dataset único

* El dataset contiene los datos metereológicos desde ene_2019 hasta may_2023

df_meteo = pd.DataFrame()
for file_name in file_names:
    df_current = pd.read_csv(ruta + file_name, delimiter=';')
    print(file_name, df_current.shape)
    df_meteo = pd.concat([df_meteo, df_current], ignore_index=True)
df_meteo.head(3)

## Seleccionar la estación de medida

* Elegimos la estación 8

df_meteo  = df_meteo[(df_meteo['ESTACION'] == 8)]
df_meteo.head(2)

# Seleccionar  los parámetros que se van a medir

* 81: Velocidad del viento
* 83: Temperatura
* 86: Humedad relativa
* 88: Precipitaciones


df_meteo = df_meteo[(df_meteo['MAGNITUD'] == 81) | (df_meteo['MAGNITUD'] == 82) | (df_meteo['MAGNITUD'] == 83) | (df_meteo['MAGNITUD'] == 86) | (df_meteo['MAGNITUD'] == 89)]
df_meteo.head(2)

df_meteo.info()

## Preparación de los datos de salida

data_columns = [ 'H01', 'H02', 'H03', 'H04', 'H05', 'H06', 'H07', 'H08', 'H09', 'H10', 'H11', 'H12', 'H13', 'H14', 'H15', 'H16', 'H17', 'H18', 'H19', 'H20', 'H21', 'H22', 'H23', 'H24'] # feature
veri_columns = [ 'V01', 'V02', 'V03', 'V04', 'V05', 'V06', 'V07', 'V08', 'V09', 'V10', 'V11', 'V12', 'V13', 'V14', 'V15', 'V16', 'V17', 'V18', 'V19', 'V20', 'V21', 'V22', 'V23', 'V24']

hora_column = np.arange(1,25) # feature label

data = pd.DataFrame(columns=['HORA', 'DATO', 'VERIFICADO'])
aux = pd.DataFrame(columns=['ANO', 'MES', 'DIA', 'HORA', 'DATO', 'VERIFICADO', 'ESTACION']) # aux de próxima fila a ser añadida
for  (h,d,v) in zip(hora_column, data_columns, veri_columns): 
    aux
    aux['DATO'] = df_meteo[ data_columns[h-1] ] 
    aux['VERIFICADO'] = df_meteo[ veri_columns[h-1] ] 
    aux['HORA'] = h-1
    aux[['ANO', 'MES', 'DIA']] = df_meteo[['ANO','MES','DIA']]
    aux['MAGNITUD'] = df_meteo ['MAGNITUD'].astype(int)
    aux['ESTACION'] = df_meteo ['ESTACION']
    data = pd.concat([data, aux], ignore_index=True)
data = data.astype({'HORA': 'int32'}, copy = False)
data.head(2)

# Añadir columna FECHA con formato datetime

data['FECHA'] = pd.to_datetime(data[['ANO','MES','DIA','HORA']].rename(columns = {'ANO': 'year', 'MES': 'month','DIA': 'day', 'HORA': 'hour'}))
data.head(2)

## Ordenar datos por orden cronológico (FECHA / HORA)

data = data.sort_values(['FECHA','HORA'])
data.head(2)

## Marcar datos no válidos en el DataFrame

Datos no verificados

data.loc[data['VERIFICADO'] != 'V', ['DATO']] = np.nan
data.loc[data['VERIFICADO'] != 'V'].head(2)

# Borrar columna verificado

data = data.drop('VERIFICADO', axis=1)
data.info()

## Añadir columnas que faltan y formatearlas

data = data.set_index(['FECHA','MAGNITUD'])['DATO'].unstack(level=-1).reset_index().rename_axis(None)
data.head(2)

## Renombrar columnas

data = data.rename(columns={83: 'TEMPERATURA', 86: 'HUMEDAD'}, copy = False, errors = 'raise')
data.head(2)

## Datos no válidos

### Comprobar el porcentaje de datos nulos

calc_missing(data)

### Comprobar que no hay valores negativos en la columna HUMEDAD

* La HUMEDAD se mide en porcentajes. por tanto, un valor negativo en esta columna indica que la medición es incorrecta.

* Si se encuentra un valor negativo en esta columna se rellenará con NaN

variables = ['TEMPERATURA', 'HUMEDAD']
datos = data [variables]

# buscamos si hay algún caso con algún campo con valor estrictamente negativo donde no debería
(datos < 0).any()

#### Buscar si hay algún caso con algún campo con valor estrictamente negativo 

humedad_neg_idx = data['HUMEDAD'] < 0
data[humedad_neg_idx]

#### Rellenar con NaN los valores negativos 

data.loc[humedad_neg_idx, 'HUMEDAD'] = np.nan
data.loc[humedad_neg_idx, 'TEMPERATURA'] = np.nan
data[humedad_neg_idx]

data.isnull().sum()

### Valores nulos

* Comprobar el porecentaje de nulos tras rellenar los valores negativos

calc_missing(data)

## Salvar datos preprocesados


data.to_csv(ruta_nuevoCSV,index=False) # Salvado con todos los datos
