==> notebooks/1.Preparacion_dataset_aire.ipynb <==
## Importar librerías
   

import pandas as pd
import numpy as np
import os
import data_util as util

## Cargar datos


ruta = '../data/aire/'
ruta_nuevoCSV = '../data/auxiliar/aire.csv'
file_names = os.listdir(ruta)
file_names

## Crear un dataset único

* El dataset contiene los datos de los contaminantes atmoféricos desde ene_2019 hasta may_2023


df_aire = pd.DataFrame()
for file_name in file_names:
    df_current = pd.read_csv(ruta + file_name, delimiter=';')
    print(file_name, df_current.shape)
    df_aire = pd.concat([df_aire, df_current], ignore_index=True)
df_aire.head(3)

## Seleccionar datos

Nos quedamos con los parámetros que nos interesan para el análisis: 
* SO2: Dióxido de Azufre
* CO: Monóxido de Carbono
* NO2: Dióxido de Nitrógeno
* PM2.5: Partículas < 2.5 μm
* PM10: Partículas < 10 μm
* O3: Ozono
* TOL: Tolueno
* BEN: Benceno
* EBE: Etilbenceno

df_aire = df_aire[(df_aire['MAGNITUD'] == 1) | (df_aire['MAGNITUD'] == 6) | (df_aire['MAGNITUD'] == 8) 
                | (df_aire['MAGNITUD'] == 9) | (df_aire['MAGNITUD'] == 10) | (df_aire['MAGNITUD'] == 14) 
                | (df_aire['MAGNITUD'] == 20) | (df_aire['MAGNITUD'] == 30)| (df_aire['MAGNITUD'] == 35)]
df_aire.head()

## Selección de estaciones de medida

* Se selecciona la estacion de medida de código_corto = 8, porque es la única que mide todos los contaminantes atmosféricos que son necesarios para calcular el índice de calidad del aire (ICA)

df_aire  = df_aire[(df_aire['ESTACION'] == 8)]
df_aire.tail(2)

## Crear una nueva columna fecha
Con formato datetime

df_aire['FECHA'] = pd.to_datetime(df_aire[['ANO','MES','DIA']].rename(columns = {'ANO': 'YEAR', 'MES': 'MONTH', 'DIA': 'DAY'}))

## Preparación de los datos de salida:
* Extraer los datos de las medidas tomadas
* Crear dataset auxiliar con los datos extraídos y las nuevas   columnas creadas.

data_columns = [ 'H01', 'H02', 'H03', 'H04', 'H05', 'H06', 'H07', 'H08', 'H09', 'H10', 'H11', 'H12', 'H13', 'H14', 'H15', 'H16', 'H17', 'H18', 'H19', 'H20', 'H21', 'H22', 'H23', 'H24'] # feature
veri_columns = [ 'V01', 'V02', 'V03', 'V04', 'V05', 'V06', 'V07', 'V08', 'V09', 'V10', 'V11', 'V12', 'V13', 'V14', 'V15', 'V16', 'V17', 'V18', 'V19', 'V20', 'V21', 'V22', 'V23', 'V24']

hora_column = np.arange(1,25) # feature label

df = pd.DataFrame(columns=['HORA', 'DATO', 'VERIFICADO'])
aux = pd.DataFrame(columns=['ANO', 'MES', 'DIA', 'HORA', 'DATO', 'VERIFICADO', 'ESTACION']) # aux de próxima fila a ser añadida
for  (h,d,v) in zip(hora_column, data_columns, veri_columns): 
    aux
    aux['DATO'] = df_aire [ data_columns[h-1] ]
    aux['VERIFICADO'] = df_aire [ veri_columns[h-1] ] 
    aux['HORA'] = h-1
    aux[['ANO', 'MES', 'DIA']] = df_aire[['ANO','MES','DIA']]
    aux['MAGNITUD'] = df_aire ['MAGNITUD'].astype(int)
    aux['ESTACION'] = df_aire ['ESTACION']
    df = pd.concat([df, aux], ignore_index=True)
df = df.astype({'HORA': 'int32'}, copy = False)
print(df.shape)
df.head(2)


## Añadir la columna fecha en formato datetime al nuevo dataset 

df['FECHA'] = pd.to_datetime(df[['ANO','MES','DIA','HORA']].rename(columns = {'ANO': 'year', 'MES': 'month','DIA': 'day', 'HORA': 'hour'}))
df.head(2)

## Ordenar datos por orden cronológico (FECHA / HORA)

df = df.sort_values(['FECHA','HORA'])
df.head(2)

## Marcar datos no válidos en el DataFrame:
* Datos no verificados

df.loc[df['VERIFICADO'] != 'V', ['DATO']] = np.nan
df.loc[df['VERIFICADO'] != 'V']

## Borrar columna verificado

df = df.drop('VERIFICADO', axis=1)
df.info()

## Añadir columnas que faltan y formatearlas,

df = df.set_index(['FECHA', 'ANO', 'MES', 'DIA', 'HORA', 'ESTACION','MAGNITUD'])['DATO'].unstack(level=-1).reset_index().rename_axis(None)
df.head()

## Renombrar columnas con los contaminantes

df = df.rename(columns={1: 'SO2', 6: 'CO', 8: 'NO2', 9: 'PM2.5', 10: 'PM10', 14: 'O3', 20: 'TOL', 30: 'BEN', 35: 'EBE'}, 
                copy = False, errors = 'raise')
df.head(2)

## Crear columna BTX

* Crear la columna BTX a partir de las columnas TOL, BEN, EBE 
+ Eliminar las columnas: TOL, BEN, EBE

df['BTX'] = df[['TOL', 'BEN', 'EBE']].sum(axis=1)
df.head(2)

df.drop(['TOL', 'BEN', 'EBE'], axis=1, inplace=True)
df.head(1)

## Añadir columna para el día de la semana

dates = df['FECHA'].dt.to_pydatetime()
days_of_week = ['Lunes', 'Martes', 'Miercoles', 'Jueves', 'Viernes', 'Sabado', 'Domingo']
df['DIA_SEMANA'] = np.array([days_of_week[d.weekday()] for d in dates])
df.head(2)

dia_semana = df.pop('DIA_SEMANA')
df.insert(5, 'DIA_SEMANA', dia_semana)
df.head(2)

df.reset_index(drop=True)
df.head(2)

## Datos no válidos

df.isnull().sum()

# Porcentaje de datos faltantes
util.calc_missing(df)

##  Salvar los datos preprocesados

Se salvan los datos tratados para su posterior fusión con otros csv.

df.to_csv(ruta_nuevoCSV, index=False)
