==> notebooks/3.Preparacion_dataset_calidad_aire.ipynb <==
## Importar librerías

import pandas as pd 
import numpy as np
import os
import data_util as util
from ydata_profiling import ProfileReport

## Cargar datos preprocesados previamente

ruta = '../data/auxiliar/aire.csv'
ruta_nuevoCSV = '../data/calidad_aire.csv'
df_aire = pd.read_csv(ruta)
df_aire.head(2)

ruta = '../data/auxiliar/meteo.csv'
df_meteo = pd.read_csv(ruta)
df_meteo.head(2)

print(df_aire.shape)
print(df_meteo.shape)

## Fusionar Dataframe

* Mediante la función merge(), se fusionan los dataframe tomando como clave para la fusión la columna 'FECHA'

# Realizar la concatenación utilizando la columna 'fecha' como clave
df_cali_aire = pd.merge(df_aire, df_meteo, on='FECHA')

# Mostrar el DataFrame concatenado
df_cali_aire.head(2)

df_cali_aire.columns

## Valores faltantes

df_cali_aire.isnull().sum()

util.calc_missing(df_cali_aire)

## Salvar datos

df_cali_aire.to_csv(ruta_nuevoCSV, index=False) # Salvado con todos los datos
